refactor CarName 객체 생성
