+++ src/main/java/racing/Car.java
@@ -6,24 +6,16 @@ import java.util.List;
 public class Car {
 
     public static final int START_POSITION = 1;
-    public static final int MAX_NAME_LENGTH = 5;
     private int position;
-    private final String name;
+    private final CarName name;
 
-    private Car(String name) {
+    private Car(CarName name) {
         this.position = START_POSITION;
         this.name = name;
     }
 
     public static Car of(String name) {
-        if (checkNameLengthExceed(name)){
-            return new Car(name);
-        }
-        throw new IllegalArgumentException("차 이름은 5자 이하여야 합니다. input: " + name);
-    }
-
-    private static boolean checkNameLengthExceed(String name) {
-        return name.length() <= MAX_NAME_LENGTH;
+        return new Car(new CarName(name));
     }
 
     public int move(boolean canProgress) {
@@ -34,7 +26,7 @@ public class Car {
     }
 
     public String getName(){
-        return this.name;
+        return this.name.getName();
     }
 
     public static List<Car> createCars(String carNames) {
