+++ src/main/java/racing/CarName.java
@@ -1,0 +1,22 @@
+package racing;
+
+public class CarName {
+    public static final int MAX_NAME_LENGTH = 5;
+    private final String name;
+
+    public CarName(String name) {
+        if (checkNameLengthExceed(name)){
+            throw new IllegalArgumentException("차 이름은 5자 이하여야 합니다. input: " + name);
+        }
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    private static boolean checkNameLengthExceed(String name) {
+        return name.length() > MAX_NAME_LENGTH;
+    }
+
+}
